LayoutTests/platform/mac-wk1/imported/w3c/web-platform-tests/mathml/presentation-markup/mpadded/mpadded-002-expected.txt


PASS mpadded with no attributes 
PASS Different widths 
PASS Different heights 
FAIL Percentage calculation for width, height and depth assert_equals: width expected 100 but got 50
PASS Different depths 
PASS Various combinations of height, depth and width. 
PASS Preferred width 
FAIL dynamic attributes (remove) assert_approx_equals: expected 100 +/- 1 but got 10
FAIL dynamic attributes (attach) assert_approx_equals: expected 50 +/- 1 but got 100
FAIL dynamic attributes (modify) assert_approx_equals: expected 50 +/- 1 but got 100
 

 

 
